backend/requirements.txt
Django==3.1.4
channels==3.0.2
channels-redis==3.2.0
django-cors-headers==3.6.0
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
django-environ==0.4.5
fastapi==0.63.0
Pillow==6.1.0
psycopg2-binary==2.8.3
uvicorn==0.11.5

